requirements.txt
numpy
pandas
Pillow
matplotlib
fontTools
tqdm